+++ src/tx_deflate.c
@@ -238,7 +238,8 @@ static void deflate_service(gpointer data)
 	if (dbg > 9)
 		printf("deflate_service: (%s) (buffer #%d, %d bytes held)\n",
 			node_ip(tx->node), attr->send_idx,
-			attr->buf[attr->send_idx].wptr - attr->buf[attr->send_idx].rptr);
+			(gint) (attr->buf[attr->send_idx].wptr - 
+					attr->buf[attr->send_idx].rptr));
 
 	/*
 	 * First, attempt to transmit the whole send buffer.
@@ -263,7 +264,7 @@ static void deflate_service(gpointer data)
 	if (b->wptr >= b->end) {
 		if (dbg > 9)
 			printf("deflate_service: (%s) sending fill buffer #%d, %d bytes\n",
-				node_ip(tx->node), attr->fill_idx, b->wptr - b->rptr);
+				node_ip(tx->node), attr->fill_idx, (gint) (b->wptr - b->rptr));
 
 		deflate_rotate_and_send(tx);
 	}
